Recursion Support
-Recursivly running on all sub-directories is now supported (automatic)
-Output option removed - all output files created in input directory

+++ AVIFrameReader.cs
@@ -52,7 +52,7 @@ namespace whiterabbitc
             public int dwEditCount;
             public int dwFormatChangeCount;
             [MarshalAs(UnmanagedType.ByValArray, SizeConst = 64)]
-            public ushort[] szName;
+            public byte[] szName;
         }
 
         private readonly IntPtr avi;
